Move modifier requirements from recipes to modifier JSON
Modifier recipe levels now use IntRange with a min and max (or an exact match via an integer). The min level represents the min level *after* applying the recipe, e.g. a min level of 2 when adding 1 level means you must have level 1 to apply this recipe.
Due to how the serializers work, all modifiers with a max level of 1 now use the compact form of setting the level to 1, max level of 2+ will use the object form.

While the requirements just run off the previously existing validation hook, to get them to show in JEI we make use of the new requirements hook, lets you add a component requirement message and a list of modifiers for display tools
Since the new system runs on modifier predicates and tool predicates, ModfierMatch has been removed. Predicates are a bit more bulky, but way more flexable.
Finally, this commit migrates modifier recipes to loadables, as it was easier to use loadables than to update the old serializers. Only reason I did not migrate earlier is I did not want to write a modifier match loadable as I knew it was getting deleted