Migrate tool parts to tool modules
As part of this migration, stat providers are now controlled by the tool definition instead of being registered with the definition. Means most of our tools are now fully interchangable, and if you wish you could just ditch tool parts on any of them.
Also means during datagen tools don't know if they are multipart or not, so we just always add render materials to render tools in datagen

This is implemented via two different hooks:
* MaterialStatsToolHook: handles building stats from the tool stat providers
* ToolPartsHook: handles listing parts for things like part swapping and tool repair
Worth considering pulling tool parts out of the definition entirely and into the recipe JSON, though that makes recycling and part swapping harder.

As part of this, the tool defintion "isMultipart" getter was replaced with "hasMaterials" as I am pushing the separation between material tools and tools with tool parts, may be useful for ancient tools if those ever become a thing
Tool definitions are also looking pretty bare at this point, will ditch the builder once I migrate the randomized materials to a module.
Tests are updated to be equivelent to the old test, though the value of a lot of them is a lot more questionable as loadables deals with the weird stuff automatically