Rework armor repair
Instead of repairing from a constant kit value, repair based on the armor max durability like vanilla does. This means for example it takes 3 ingots of cobalt to repair any piece of plate armor, instead of that varying based on the pieces durability.
This is accomplished by letting you set the max durability in the repair module in the tool definition as an alternative to the stat type.
As a result of this change, we no longer need repair kit stats to define an int value, so that stat type is now a "statless" stat type (still needed to ensure weird repair kits like leather exist despite having no heads)
Also as part of this, made the tool module builder more generic, now it lives in armor slot type