Let modifier recipes decide whether to check the total level or just the recipe level
Modifiers such as pathing, bouncy, etc. don't want to be added to a tool that starts with it as a trait.
While implementing this feature, discovered that most modifiers were accidently checking the total level instead of the recipe level (causing wrong behavior for things that are also traits like smite). The default for the new behavior has been set to recipe modifiers only to restore the 1.18 behavior