Fix tank modifier models discarding the non-fluid and offsetting weirdly (#5227)
Turns out the method of using the model bakery applies an origin to the transformation, meaning we get a funny offset with the fluid quads. Thus, we just reverse that before calling the method
No idea why this is not needed in the fluid container model, perhaps we just do no scaling there so it does not end up mattering?